Refactor: #82 소셜 로그인 관련 클래스 유저 패키지에 통합
