--- src/main/java/org/bbagisix/oauth2/dto/UserDTO.java
@@ -1,13 +1,0 @@
-package org.bbagisix.oauth2.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserDTO {
-	private String role;
-	private String name;
-	private String nickname;
-	private String email;
-}
