+++ src/main/java/org/bbagisix/user/controller/OAuth2Controller.java
@@ -1,6 +1,6 @@
-package org.bbagisix.oauth2;
+package org.bbagisix.user.controller;
 
-import org.bbagisix.oauth2.service.OAuth2Service;
+import org.bbagisix.user.service.OAuth2Service;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
