+++ src/main/java/org/bbagisix/user/dto/UserDTO.java
@@ -1,5 +1,13 @@
 package org.bbagisix.user.dto;
 
-public class UserDTO {
+import lombok.Getter;
+import lombok.Setter;
 
+@Getter
+@Setter
+public class UserDTO {
+	private String role;
+	private String name;
+	private String nickname;
+	private String email;
 }
